queue debugging and fixes, time handling improvements, basic import control

+++ exception/logic/ExceptionLogLogic.java
@@ -13,7 +13,7 @@ interface ExceptionLogLogic {
 			String source,
 			String summary,
 			String dump,
-			Optional<Integer> userId,
+			Optional<Long> userId,
 			GenericExceptionResolution resolution);
 
 }
